Ruleset: allow for usage as a standard
In contrast to plain ruleset, a PHPCS "standard" can:
* Be registered with PHPCS and used by name, both on the command line as well as in custom rulesets.
* Contain custom sniffs, if so desired.

For a ruleset to be recognized as a standard by PHPCS, it should:
* Live in a folder which matches the name of the standard
* That folder must contain a `ruleset.xml` file.

By moving the `ruleset.xml` file to a `PHPParallelLint` folder and changing the name of the standard to `PHPParallelLint`, this ruleset should now be usable as a standard.

Includes updating the description to match the current reality.