Make the PHPCS standard self-contained
As things were, this package contained a ruleset, but did not declare its dependency on PHP_CodeSniffer, so the package would not be usable without the dependant declaring a dependency on PHPCS.

This commit fixes this by:
1. Declaring PHPCS as a dependency in the `composer.json` file.
2. Declaring a dependency on the DealerDirect Composer plugin which can automatically handle the registration of PHPCS standards with PHPCS.

With these two dependencies in place, the package can be used as a self-contained package.

Includes adding a `.gitignore` file to prevent the `composer.lock` and `vendor` directory from being committed, as well as a `.gitattributes` file to prevent the `.gitignore` (and `.gitattributes`) file from being included in distributed packages.