Composer/Ruleset: include the PHPCompatibility standard
The PHPCompatibility standard can detect a large range of PHP cross-version incompatibilities and flag these, to prevent code being committed to the projects which is incompatible with the minimum supported PHP version of those projects.

For optimal results, the projects using this standard should set a `testVersion` to inform PHPCompatibility which PHP versions the code should be compatible with.

For more information, see the README of the project.

Ref: https://github.com/PHPCompatibility/PHPCompatibility/